widgetype not loading on initial render because widgettype is not set in userDB please do not change the way images are currently rendered.
widgets are loaded from widget DB when you "add a widget"
widgets are loaded from user DB when you initially render dashbaord.

We need to make it so that widgets always load from widget DB and never
from userdb as it does now. The user DB should only contain a collection
of UniqueIDs (meaning the widgets they added to account) and then using
that unique ID it then loads that widget from WidgetDB

For widgets that need a background image, currently we're loading the
images based on their ID. So images are named by ID after the widget the
represent. These do not need to be renamed unless the uniqueID's are
changed. Since we don't need to reset DB anymore we shouldn't need to
change imaage names. A script is in progress to help with renaming our
images when we have a lof of widgets

listWidget.js
the widget IDs are only hardcoded because we don't have a script to auto
them from the database yet. (I'll have it done eventually-but DB won't
be rest so thses ID shouldn't change, so this script isn't priority)