Notepad turned into widget
User can add and remove as a normal widget.
to make it default for all new users in DB add the notepad widget ID and
it should load it by default but the user should be able to remove it.

Rain please continue working on Widget Data Storage, and prove that it
works through the notepad widget.